#103 direct exec ui.on_startup if already started
